backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.2
qdrant-client==1.7.0
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
tiktoken==0.5.2
markdown==3.5.1
PyYAML==6.0.1
beautifulsoup4==4.12.2